requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
aiogram==3.4.0
python-dotenv==1.0.0
httpx==0.24.1
SQLAlchemy==2.0.20
psycopg[binary]==3.2.3
alembic==1.11.1
argon2-cffi==23.1.0
PyJWT==2.9.0
jinja2==3.1.4
aiofiles==23.1.0
python-multipart==0.0.6
redis==5.0.10
